+++ src/main/scala/onion/compiler/SemanticErrorReporter.scala
@@ -382,6 +382,10 @@ class SemanticErrorReporter(threshold: Int) {
       "error.semantic.nullableMemberAccess",
       Seq(items => asString(items(0)))
     ),
+    SemanticError.STATIC_CALL_ON_INSTANCE -> ErrorDef(
+      "error.semantic.staticCallOnInstance",
+      Seq(items => asString(items(0)))
+    ),
     SemanticError.LABEL_NOT_FOUND -> ErrorDef(
       "error.semantic.labelNotFound",
       Seq(items => asString(items(0)))